Move antlr gen js code to a seperate folder
